Fixed Error with Building Graph
Excepted a KeyError in build_graph becuase the given KeyError was caused
by the package not being listed in pip freeze. I made the decision not
to add a node for that package to the graph becuase it is either a
default pip package and should not be touched, or it is missing and will
be left for pip to handle. May potentially implement method to warn the
user in the future.